pyproject.toml
[project]
name = "lennys-wisdom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ManevartheGIT", email = "anantha.git@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.0",
    "httpx>=0.28.1",
    "markdownify>=1.2.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-frontmatter>=1.1.0",
    "sqlite-utils>=3.39",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"
